payment booking

+++ app/helpers/sessions_helper.rb
@@ -34,6 +34,7 @@ module SessionsHelper
 
     session[:tickets_id].each do |ticket_id|
       ticket = Ticket.find_by(id: ticket_id)
+      next if ticket.blank?
       ticket.seat.destroy
     end
     session.delete(:tickets_id)
@@ -46,5 +47,44 @@ module SessionsHelper
       @seats << ticket.seat.seat_number
     end
     @seats
+  end
+
+  def load_payments
+    tickets = []
+    return if session[:tickets_id].blank?
+
+    session[:tickets_id].each do |ticket_id|
+      ticket = Ticket.find_by(id: ticket_id)
+      tickets << ticket
+    end
+    tickets
+  end
+
+  def save_payment id
+    Payment.transaction do
+      total_cost = session[:tickets_id].length * Settings.price.standard
+      payment = Payment.find_by(id: id)
+      payment.update(status: :approve,
+                     total_cost: total_cost)
+    end
+  end
+
+  def save_ticket id
+    save_payment id
+    return false if session[:tickets_id].blank?
+
+    ticket_transaction = Ticket.transaction do
+      session[:tickets_id].each do |ticket_id|
+        ticket = Ticket.find_by(id: ticket_id)
+        next unless ticket.update(payment_id: id)
+        ticket.seat.unavailable!
+      end
+      true
+    rescue ActiveRecord::Rollback
+      false
+    end
+
+    session[:tickets_id] = []
+    ticket_transaction
   end
 end
